Add files via upload
This Python code is an implementation of the attached algorithm. 
The authors in this work propose an original fault signature based on an improved combination of Hilbert and Park
transforms. Starting from this combination we can release two fault signatures: Hilbert modulus current
space vector (HMCSV) and Hilbert phase current space vector (HPCSV). These two signatures are subsequently analyzed using the classical fast Fourier transform (FFT). The effects of HMCSV and HPCSV spectrums are described and the related frequencies are determined. A comparative study is presented of the
suggested signature (HPCSV) and the MCSA which is the signature more recently proposed in the literature. The proposed signature shows its effectiveness and its robustness in both electrical and mechanical
fault detection. The magnitudes of spectral components relative to the studied faults are extracted in
order to develop the input vector necessary for the pattern recognition tool based on support vector
machine (SVM) approach with an aim of classifying automatically the various states of the induction
motor. This approach was applied to a 1.1 kw induction motor under normal operation and with the following faults: unbalanced voltage, broken rotor bar, air-gap eccentricity and outer raceway ball bearing
defect